requirements.txt
discord.py>=2.3.2
sqlalchemy>=2.0.23
asyncpg>=0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
Pillow>=10.1.0
python-dateutil>=2.8.2
alembic>=1.13.1

